Discord-bot/requirements.txt
# requirements.txt
discord.py>=2.3.0
yt-dlp>=2023.12.30
PyNaCl>=1.4.0  # Required for voice support
sqlite3  # Usually included with Python
python-dateutil>=2.8.2  # For date handling
asyncio>=3.4.3